feat: standar with integer in user class and test regsiter method
In order to create new users

+++ src/main/java/netlife/devmasters/booking/service/UserService.java
@@ -17,7 +17,7 @@ public interface UserService {
 	User registrar(User usuario)
 			throws UsuarioNoEncontradoExcepcion, NombreUsuarioExisteExcepcion, EmailExisteExcepcion, MessagingException, IOException;
 
-	public Optional<User> getById(Long codigo);
+	public Optional<User> getById(Integer codigo);
 
 	List<User> getUsuarios();
